add index-nosvg.html and format_html.pl
This is a basic page to show off the images.  The source file just using
plain PNG images, and from this I build an SVG version.  Firefox does
not render the SVG at the appropriate size, so I have to look at the PNG
and determine the size manually.  Also, I use an embed tag, instead of
an iframe or object tag, since it seems to work better on Firefox 3.6.
I'm not sure about other browsers.