publish: make more robust with plumbing commands
Using git-checkout had the disadvantage that it would delete the
committed files.  Instead, use git-symbolic-ref to manually change
branches, after first verifying that the work tree and index are clean.
Then, delete the index and add the files as before.  However, only
commit if there is a change.  Finally, switch back using
git-symbolic-ref.