publish: back up and restore old index; use add -f
Instead of deleting the index at the start and forgetting to restore it
at the end, we now rename the index file and restore this backup upon
completion.

The git-add command fails if a named file is in .gitignore.  In our
case, most of the files are, so we need the -f flag to skip the ignore
file.