fix direction of diff arrows
As suggested by Jon Seymour, fix the direction of the diffs.  The D->B
diff was not consistent with the rest, but that was the only one that
made sense.  Now, all of the other arrows have been reversed.  This
shows that the A->B is from (-) A to (+) B.