clarify detached HEAD, and move it up higher
The detached HEAD state comes up frequently, so instead of hiding it at
the bottom, move it up to the checkout section.  Add a note about why
you might want to do this, and also add a new graph for 'checkout -b'.

Thanks to Daniel Barkalow for the suggestion (though I'm not yet sure if
this is what he had in mind.)