perform automatic detection of SVG capabilities
We now only have a single page, and the detection of SVG vs PNG is done
in Javascript.  The default is to display PNG, but if the scripting is
enabled and SVG capability is detected, all of the PNG images are
replaced by SVG.  This can be disabled by putting "no-svg" in the
query string.  Additionally, the display of SVG images is fixed on
Webkit (Chrome, Safari, etc.)

Webkit supports SVG img elements (and, in fact, does not display SVG
object elements correctly when zooming), so on these browsers we can
just change the img's src attribute.  Firefox does not support SVG img
elements, so we have to replace these elements with an object element of
the same size.

To simplify the detection of PNG images with an SVG equivalent, rename
all of the PNGs to ".svg.png".

Add a symlink from index-svg.html to index.html for folks with old
bookmarks.