fix arrow heads for edges
The 'edge' command does not actually create a path, so for all the cases
where we were creating a bunch of edges, the path was actually just
a single point.  For some reason, TikZ adds an arrowhead if the path
style contains ->, even though no actual path was created.  That is, an
arrowhead was created wherever the last edge started, which is very
annoying.  So, to get around this, set the 'every edge' style to be ->,
but not the 'line' style.