Use a scaling factor for all sizes
Instead of hardcoding all the units in the styles, use \g in place of
cm, \gLine in place of pt, and \gFont in place of \Large.  Now, it is
easy to scale the size of the images by changing these three parameters.
This scaling is not perfect, and you're limited to increments of the
font size, but it's still a heck of a lot easier than adjusting all the
dimensions manually.

Thanks for Thomas Koch for the idea.