Turn on Google Analytics.
A long time ago, I turned off Google Analytics in the master branch and
turned in on in the "production" branch so that people who fork my code
wouldn't have the analytics code running on their fork.  The problem is
that I never remember to push from the production branch, so in effect
I have had Analytics turned off for years now!  So, the heck with forks,
I'm turning this on by default.